rename datasource file
