+++ src/components/QueryEditor/XrayLinks.tsx
@@ -1,6 +1,6 @@
 import { useTheme2 } from '@grafana/ui';
 import { css } from '@emotion/css';
-import { XrayDataSource } from '../../DataSource';
+import { XrayDataSource } from '../../XRayDataSource';
 import { XrayQuery } from '../../types';
 import { GrafanaTheme2, TimeRange } from '@grafana/data';
 import React from 'react';
