+++ src/module.ts
@@ -3,7 +3,7 @@ import { DataSourcePlugin } from '@grafana/data';
 import CheatSheet from './components/CheatSheet';
 import { ConfigEditor } from './components/ConfigEditor/ConfigEditor';
 import { QueryEditor } from './components/QueryEditor/QueryEditor';
-import { XrayDataSource } from './DataSource';
+import { XrayDataSource } from './XRayDataSource';
 import { XrayJsonData, XrayQuery } from './types';
 
 export const plugin = new DataSourcePlugin<XrayDataSource, XrayQuery, XrayJsonData, AwsAuthDataSourceSecureJsonData>(
